Made improvements to functions and seperated related functions into modules
