+++ src/firebase/appConfig.js
@@ -1,0 +1,27 @@
+import firebase from 'firebase/compat/app';
+import 'firebase/compat/firestore';
+import 'firebase/compat/auth';
+
+/**
+ * Initializes a Firebase App and the required services.
+ * @returns Firebase App
+ */
+function firebaseApp() {
+    const app = firebase.initializeApp({
+        apiKey: process.env.REACT_APP_API_KEY,
+        authDomain: process.env.REACT_APP_AUTH_DOMAIN,
+        projectId: process.env.REACT_APP_PROJECT_ID,
+        storageBucket: process.env.REACT_APP_STORAGE_BUCKET,
+        messagingSenderId: process.env.REACT_APP_MSG_SENDER_ID,
+        appId: process.env.REACT_APP_APP_ID,
+        measurementId: process.env.REACT_APP_MEASUREMENT_ID
+    });
+    const auth = firebase.auth();
+    const firestore = firebase.firestore();
+
+    return {
+        app, auth, firestore
+    }
+}
+
+export default firebaseApp;
